# storage/api-docs/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Company Fleet API",
        "description": "API для служебных автомобилей",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Local dev server"
        }
    ],
    "paths": {
        "/api/ping": {
            "get": {
                "summary": "Check API ping",
                "operationId": "733e904e1c9d2c540abcb3a716b2980e",
                "responses": {
                    "200": {
                        "description": "API is alive"
                    }
                }
            }
        },
        "/api/tasks": {
            "get": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Получить список задач",
                "description": "Возвращает список задач с фильтрами по статусу, приоритету и пользователю. Сортировка по дате создания (новые первыми).",
                "operationId": "getTasks",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Фильтр по статусу задачи",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "new",
                                "in_progress",
                                "completed",
                                "cancelled"
                            ]
                        }
                    },
                    {
                        "name": "priority",
                        "in": "query",
                        "description": "Фильтр по приоритету задачи",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "high",
                                "normal",
                                "low"
                            ]
                        }
                    },
                    {
                        "name": "user_id",
                        "in": "query",
                        "description": "Фильтр по ID пользователя",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список задач",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Task"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Некорректный запрос"
                    }
                }
            },
            "post": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Создать новую задачу",
                "description": "Создание задачи с обязательным полем title. Если user_id не указан или пользователь не найден, назначаем менеджера. Статус по умолчанию = 'new', если priority = 'high' → status = 'in_progress'.",
                "operationId": "createTask",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Fix login bug"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Login fails on Safari"
                                    },
                                    "user_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "priority": {
                                        "type": "string",
                                        "enum": [
                                            "high",
                                            "normal",
                                            "low"
                                        ],
                                        "example": "high"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Задача успешно создана",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Task"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации"
                    }
                }
            }
        },
        "/api/tasks/{id}/status": {
            "put": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Изменить статус задачи",
                "description": "Изменение статуса задачи. При смене статуса на completed автоматически добавляется комментарий. Также запускается job уведомлений для всех менеджеров.",
                "operationId": "updateTaskStatus",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID задачи",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "status"
                                ],
                                "properties": {
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "new",
                                            "in_progress",
                                            "completed",
                                            "cancelled"
                                        ],
                                        "example": "completed"
                                    },
                                    "user_id": {
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Статус задачи обновлён",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Task"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Задача не найдена"
                    },
                    "422": {
                        "description": "Ошибка валидации"
                    }
                }
            }
        },
        "/api/tasks/{id}/comments": {
            "post": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Добавить комментарий к задаче",
                "description": "Добавление комментария к задаче. Поля: comment (обязательно), user_id (обязательно). Нельзя добавлять комментарии к cancelled задачам.",
                "operationId": "addComment",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID задачи",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "comment",
                                    "user_id"
                                ],
                                "properties": {
                                    "comment": {
                                        "type": "string",
                                        "example": "This is a comment"
                                    },
                                    "user_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Комментарий успешно создан"
                    },
                    "403": {
                        "description": "Нельзя добавлять комментарии к cancelled задачам"
                    },
                    "404": {
                        "description": "Задача не найдена"
                    },
                    "422": {
                        "description": "Ошибка валидации"
                    }
                }
            }
        },
        "/api/tasks/{id}": {
            "get": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Получить задачу с комментариями",
                "description": "Возвращает задачу по ID вместе с комментариями и информацией о пользователях (name, position)",
                "operationId": "getTask",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID задачи",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Задача успешно получена"
                    },
                    "404": {
                        "description": "Задача не найдена"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Task": {
                "title": "Task",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "title": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "user_id": {
                        "type": "integer"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "new",
                            "in_progress",
                            "completed",
                            "cancelled"
                        ]
                    },
                    "priority": {
                        "type": "string",
                        "enum": [
                            "high",
                            "normal",
                            "low"
                        ]
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Tasks",
            "description": "Tasks"
        }
    ]
}